starting roadrunner installation
