+++ build.dependencies.gradle
@@ -22,4 +22,3 @@ dependencies {
 
     implementation 'com.acmerobotics.dashboard:dashboard:0.4.5'
 }
-
